Conflicts:
	scilab/Makefile.in
	scilab/aclocal.m4
	scilab/configure
	scilab/contrib/Makefile.in
	scilab/desktop/Makefile.in
	scilab/desktop/images/icons/Makefile.in
	scilab/modules/Makefile.in
	scilab/modules/action_binding/Makefile.in
	scilab/modules/api_scilab/Makefile.in
	scilab/modules/arnoldi/Makefile.in
	scilab/modules/ast/Makefile.in
	scilab/modules/ast/src/cpp/ast/run_AssignExp.hpp
	scilab/modules/atoms/Makefile.in
	scilab/modules/boolean/Makefile.in
	scilab/modules/cacsd/Makefile.in
	scilab/modules/call_scilab/Makefile.in
	scilab/modules/commons/Makefile.in
	scilab/modules/compatibility_functions/Makefile.in
	scilab/modules/completion/Makefile.in
	scilab/modules/console/Makefile.in
	scilab/modules/core/Makefile.in
	scilab/modules/core/sci_gateway/cpp/sci_error.cpp
	scilab/modules/coverage/Makefile.in
	scilab/modules/data_structures/Makefile.in
	scilab/modules/demo_tools/Makefile.in
	scilab/modules/development_tools/Makefile.in
	scilab/modules/development_tools/src/fake/Makefile.in
	scilab/modules/differential_equations/Makefile.in
	scilab/modules/dynamic_link/Makefile.in
	scilab/modules/elementary_functions/Makefile.am
	scilab/modules/elementary_functions/Makefile.in
	scilab/modules/elementary_functions/macros/permute.sci
	scilab/modules/elementary_functions/sci_gateway/cpp/elem_func_gw/elem_func_gw.vcxproj
	scilab/modules/external_objects/Makefile.in
	scilab/modules/external_objects_java/Makefile.in
	scilab/modules/fftw/Makefile.in
	scilab/modules/fileio/Makefile.in
	scilab/modules/functions/Makefile.in
	scilab/modules/functions_manager/Makefile.in
	scilab/modules/functions_manager/src/cpp/dynamic_modules.cpp
	scilab/modules/genetic_algorithms/Makefile.in
	scilab/modules/graph/Makefile.in
	scilab/modules/graphic_export/Makefile.in
	scilab/modules/graphic_objects/Makefile.in
	scilab/modules/graphics/Makefile.in
	scilab/modules/gui/Makefile.in
	scilab/modules/hdf5/Makefile.in
	scilab/modules/hdf5/src/c/hdf5.vcxproj
	scilab/modules/helptools/Makefile.in
	scilab/modules/history_browser/Makefile.in
	scilab/modules/history_manager/Makefile.in
	scilab/modules/integer/Makefile.in
	scilab/modules/interpolation/Makefile.in
	scilab/modules/io/Makefile.in
	scilab/modules/javasci/Makefile.in
	scilab/modules/jvm/Makefile.in
	scilab/modules/linear_algebra/Makefile.in
	scilab/modules/linear_algebra/tests/unit_tests/cond.dia.ref
	scilab/modules/localization/Makefile.in
	scilab/modules/m2sci/Makefile.in
	scilab/modules/matio/Makefile.in
	scilab/modules/mexlib/Makefile.in
	scilab/modules/modules_manager/Makefile.in
	scilab/modules/mpi/Makefile.in
	scilab/modules/optimization/Makefile.in
	scilab/modules/output_stream/Makefile.in
	scilab/modules/overloading/Makefile.in
	scilab/modules/parallel/Makefile.in
	scilab/modules/parameters/Makefile.in
	scilab/modules/polynomials/Makefile.in
	scilab/modules/prebuildjava/Makefile.in
	scilab/modules/preferences/Makefile.in
	scilab/modules/randlib/Makefile.in
	scilab/modules/renderer/Makefile.in
	scilab/modules/scicos/Makefile.in
	scilab/modules/scicos/sci_gateway/cpp/sci_scicos_debug.cpp
	scilab/modules/scicos_blocks/Makefile.in
	scilab/modules/scinotes/Makefile.in
	scilab/modules/signal_processing/Makefile.in
	scilab/modules/simulated_annealing/Makefile.in
	scilab/modules/slint/Makefile.in
	scilab/modules/sound/Makefile.in
	scilab/modules/sparse/Makefile.in
	scilab/modules/sparse/sci_gateway/cpp/sci_sparse.cpp
	scilab/modules/special_functions/Makefile.in
	scilab/modules/spreadsheet/Makefile.in
	scilab/modules/statistics/Makefile.in
	scilab/modules/string/Makefile.in
	scilab/modules/tclsci/Makefile.in
	scilab/modules/threads/Makefile.in
	scilab/modules/time/Makefile.in
	scilab/modules/types/Makefile.in
	scilab/modules/ui_data/Makefile.in
	scilab/modules/umfpack/Makefile.in
	scilab/modules/webtools/Makefile.in
	scilab/modules/windows_tools/Makefile.in
	scilab/modules/xcos/Makefile.in
	scilab/modules/xml/Makefile.in

+++ scilab/modules/scicos/src/cpp/view_scilab/BaseAdapter.hxx
@@ -75,7 +75,7 @@ public:
     property(const std::wstring& prop, getter_t g, setter_t s) : original_index(fields.size()), name(prop), get(g), set(s) {};
     ~property() {};
 
-    size_t original_index;
+    int original_index;
     std::wstring name;
     getter_t get;
     setter_t set;
